Move controllers into controller folder
