+++ core/database/foxx/api/utils/posix_path.js
