+++ core/database/foxx/tests/project.test.js
@@ -2,7 +2,7 @@
 
 const chai = require("chai");
 const expect = chai.expect;
-const Project = require("../api/project");
+const Project = require("../api/controllers/project");
 const g_db = require("@arangodb").db;
 const g_lib = require("../api/support");
 const arangodb = require("@arangodb");
